refactor: split chat screen helper modules
