+++ services/aris-web/app/sessions/[sessionId]/chat-screen/helpers.tsx
@@ -1,20 +1,13 @@
-import React from 'react';
-import type { ComponentType, ReactNode } from 'react';
+import type { ComponentType } from 'react';
 import {
   Brain,
   CircleAlert,
-  Cpu,
-  File,
-  FileCode,
   FilePenLine,
   FileSearch,
-  FileText,
-  Folder,
   FolderTree,
   MessageSquareText,
   TerminalSquare,
 } from 'lucide-react';
-import { readLocalStorage, writeLocalStorage } from '@/lib/browser/localStorage';
 import {
   isRunLifecycleEvent,
   isTerminalRunStatus,
@@ -22,7 +15,6 @@ import {
   readUiEventStreamEvent,
 } from '@/lib/happy/chatRuntime';
 import type {
-  AgentFlavor,
   ApprovalPolicy,
   ChatImageAttachment,
   SessionChat,
@@ -30,44 +22,22 @@ import type {
   UiEventKind,
   UiEventResult,
 } from '@/lib/happy/types';
-import {
-  DEFAULT_GEMINI_MODE_ID,
-  GEMINI_MODE_SELECTION_OPTIONS,
-  deriveOpenAiModelLabel,
-  type ProviderModelSelections,
-} from '@/lib/settings/providerModels';
-import {
-  ClaudeIcon,
-  CodexIcon,
-  DockerLogoIcon,
-  GeminiIcon,
-  GitLogoIcon,
-} from '@/components/ui/AgentIcons';
+import { DockerLogoIcon, GitLogoIcon } from '@/components/ui/AgentIcons';
 import {
   ACTION_COLLAPSE_THRESHOLD,
   CHAT_SIDEBAR_SECTION_LABELS,
-  COMPOSER_MODELS_BY_AGENT,
   DEFAULT_CHAT_RUNTIME_UI_STATE,
   PREVIEW_MAX_CHARS,
   PREVIEW_MAX_LINES,
-  RECENT_FILES_MAX,
-  RECENT_FILES_STORAGE_KEY,
-  WORKSPACE_FILE_OPEN_EVENT,
 } from './constants';
 import type {
   ActionKind,
-  AgentMeta,
   ChatSidebarSnapshot,
-  ComposerModelOption,
   ContextItem,
   FolderLabel,
-  GeminiModeOption,
-  LegacyCustomModels,
-  ModelReasoningEffort,
   ResourceLabel,
   StreamRenderItem,
   Tone,
-  WorkspaceFileOpenDetail,
 } from './types';
 import { FOLDER_LABELS } from './types';
 import {
@@ -75,64 +45,34 @@ import {
   parseLocalFileReferenceTarget,
   scanMarkdownLinks,
 } from '../chatFileReferences';
-
-export function getRecentFiles(): string[] {
-  try {
-    return JSON.parse(readLocalStorage(RECENT_FILES_STORAGE_KEY) ?? '[]') as string[];
-  } catch {
-    return [];
-  }
-}
-
-export function saveRecentFile(filePath: string): void {
-  try {
-    const prev = getRecentFiles().filter((path) => path !== filePath);
-    writeLocalStorage(RECENT_FILES_STORAGE_KEY, JSON.stringify([filePath, ...prev].slice(0, RECENT_FILES_MAX)));
-  } catch {
-    // localStorage unavailable
-  }
-}
-
-export async function copyTextToClipboard(text: string): Promise<void> {
-  if (typeof navigator !== 'undefined' && navigator.clipboard?.writeText) {
-    await navigator.clipboard.writeText(text);
-    return;
-  }
-
-  if (typeof document === 'undefined') {
-    throw new Error('clipboard-unavailable');
-  }
-
-  const textarea = document.createElement('textarea');
-  textarea.value = text;
-  textarea.setAttribute('readonly', 'true');
-  textarea.style.position = 'fixed';
-  textarea.style.top = '0';
-  textarea.style.left = '0';
-  textarea.style.opacity = '0';
-  document.body.appendChild(textarea);
-  textarea.focus();
-  textarea.select();
-
-  const copied = document.execCommand('copy');
-  document.body.removeChild(textarea);
-
-  if (!copied) {
-    throw new Error('clipboard-unavailable');
-  }
-}
-
-export function getFileIcon(name: string, isDirectory: boolean): ReactNode {
-  if (isDirectory) return <Folder size={14} />;
-  const ext = name.split('.').pop()?.toLowerCase() ?? '';
-  if (['ts', 'tsx', 'js', 'jsx', 'py', 'go', 'rs', 'java', 'c', 'cpp', 'cs'].includes(ext)) {
-    return <FileCode size={14} />;
-  }
-  if (['md', 'txt', 'yaml', 'yml', 'toml', 'json'].includes(ext)) {
-    return <FileText size={14} />;
-  }
-  return <File size={14} />;
-}
+export {
+  buildChatUrl,
+  copyTextToClipboard,
+  dispatchWorkspaceFileOpen,
+  getFileIcon,
+  getRecentFiles,
+  isWorkspacePathWithinRoot,
+  joinWorkspacePath,
+  normalizeWorkspaceClientPath,
+  readChatIdFromLocation,
+  saveRecentFile,
+  writeChatIdToHistory,
+} from './browserHelpers';
+export {
+  deriveGeminiModeLabel,
+  normalizeAgentFlavor,
+  normalizeGeminiModeId,
+  normalizeModelId,
+  normalizeModelReasoningEffort,
+  resolveAgentMeta,
+  resolveAgentSubtitle,
+  resolveAvailableComposerModelId,
+  resolveAvailableGeminiModeId,
+  resolveComposerModels,
+  resolveDefaultGeminiModeId,
+  resolveDefaultModelId,
+  resolveGeminiModeOptions,
+} from './modelHelpers';
 
 export function extractImageAttachments(items: ContextItem[]): ChatImageAttachment[] {
   return items.flatMap((item) => (item.type === 'image' ? [item.attachment] : []));
@@ -155,76 +95,6 @@ export function fileExtension(filename: string): string {
   const ext = base.slice(dotIndex + 1).toLowerCase();
   if (!/^[a-z0-9]+$/.test(ext)) return '';
   return ext;
-}
-
-export function normalizeWorkspaceClientPath(input: string): string {
-  const trimmed = input.trim();
-  if (!trimmed) {
-    return '/';
-  }
-
-  const withLeadingSlash = trimmed.startsWith('/') ? trimmed : `/${trimmed}`;
-  const normalized = withLeadingSlash.replace(/\/+/g, '/').replace(/\/$/, '');
-  return normalized || '/';
-}
-
-export function joinWorkspacePath(dirPath: string, name: string): string {
-  const normalizedDir = normalizeWorkspaceClientPath(dirPath);
-  const trimmedName = name.trim().replace(/^\/+/, '');
-  return normalizedDir === '/' ? `/${trimmedName}` : `${normalizedDir}/${trimmedName}`;
-}
-
-export function buildChatUrl(sessionId: string, chatId: string): string {
-  return `/sessions/${encodeURIComponent(sessionId)}?chat=${encodeURIComponent(chatId)}`;
-}
-
-export function readChatIdFromLocation(): string | null {
-  if (typeof window === 'undefined') {
-    return null;
-  }
-  const raw = new URL(window.location.href).searchParams.get('chat');
-  if (typeof raw !== 'string') {
-    return null;
-  }
-  const trimmed = raw.trim();
-  return trimmed || null;
-}
-
-export function writeChatIdToHistory(url: string, mode: 'push' | 'replace' = 'push') {
-  if (typeof window === 'undefined') {
-    return;
-  }
-  const current = `${window.location.pathname}${window.location.search}`;
-  if (current === url) {
-    return;
-  }
-  if (mode === 'replace') {
-    window.history.replaceState({}, '', url);
-    return;
-  }
-  window.history.pushState({}, '', url);
-}
-
-export function isWorkspacePathWithinRoot(targetPath: string, rootPath: string): boolean {
-  const normalizedTarget = normalizeWorkspaceClientPath(targetPath);
-  const normalizedRoot = normalizeWorkspaceClientPath(rootPath);
-  return normalizedRoot === '/'
-    ? normalizedTarget.startsWith('/')
-    : normalizedTarget === normalizedRoot || normalizedTarget.startsWith(`${normalizedRoot}/`);
-}
-
-export function dispatchWorkspaceFileOpen(detail: WorkspaceFileOpenDetail) {
-  if (typeof window === 'undefined') {
-    return;
-  }
-
-  window.dispatchEvent(new CustomEvent<WorkspaceFileOpenDetail>(WORKSPACE_FILE_OPEN_EVENT, {
-    detail: {
-      ...detail,
-      path: normalizeWorkspaceClientPath(detail.path),
-      line: typeof detail.line === 'number' ? detail.line : null,
-    },
-  }));
 }
 
 export function classifyLabelLink(label: string, rawPath: string): ResourceLabel | null {
@@ -281,240 +151,6 @@ export function classifyPath(pathValue: string): ResourceLabel | null {
   }
 
   return null;
-}
-
-export function resolveAgentMeta(agentFlavor: string): AgentMeta {
-  if (agentFlavor === 'claude') {
-    return { label: 'Claude', tone: 'clay', Icon: ClaudeIcon };
-  }
-  if (agentFlavor === 'codex') {
-    return { label: 'Codex', tone: 'mint', Icon: CodexIcon };
-  }
-  if (agentFlavor === 'gemini') {
-    return { label: 'Gemini', tone: 'blue', Icon: GeminiIcon };
-  }
-  return { label: 'Runtime', tone: 'blue', Icon: Cpu };
-}
-
-export function resolveAgentSubtitle(agentFlavor: string): string {
-  if (agentFlavor === 'claude') return '균형 잡힌 코딩 흐름';
-  if (agentFlavor === 'codex') return '빠른 구현 및 실행';
-  if (agentFlavor === 'gemini') return '넓은 맥락과 추론';
-  return '에이전틱 런타임';
-}
-
-export function normalizeAgentFlavor(value: unknown, fallback: AgentFlavor = 'codex'): AgentFlavor {
-  if (value === 'claude' || value === 'codex' || value === 'gemini') {
-    return value;
-  }
-  return fallback;
-}
-
-export function normalizeModelId(value: unknown): string | null {
-  if (typeof value !== 'string') {
-    return null;
-  }
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return null;
-  }
-  const canonical = trimmed === 'gpt-5-codex' ? 'gpt-5.3-codex' : trimmed;
-  return canonical.slice(0, 120);
-}
-
-export function normalizeGeminiModeId(value: unknown): string | null {
-  if (typeof value !== 'string') {
-    return null;
-  }
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return null;
-  }
-  return trimmed.slice(0, 120);
-}
-
-export function normalizeModelReasoningEffort(
-  value: unknown,
-  fallback: ModelReasoningEffort = 'medium',
-): ModelReasoningEffort {
-  if (value === 'low' || value === 'medium' || value === 'high' || value === 'xhigh') {
-    return value;
-  }
-  return fallback;
-}
-
-export function isSupportedAgentFlavor(value: AgentFlavor): value is 'codex' | 'claude' | 'gemini' {
-  return value === 'codex' || value === 'claude' || value === 'gemini';
-}
-
-export function deriveGeminiModeLabel(modeId: string): string {
-  const normalized = modeId.trim().toLowerCase();
-  if (normalized === 'default') {
-    return 'Default';
-  }
-  if (normalized === 'yolo') {
-    return 'YOLO';
-  }
-  if (normalized === 'plan') {
-    return 'Plan';
-  }
-  if (normalized === 'autoedit' || normalized === 'auto_edit') {
-    return 'Auto Edit';
-  }
-  return modeId
-    .replace(/[_-]+/g, ' ')
-    .replace(/\b\w/g, (char) => char.toUpperCase());
-}
-
-export function resolveGeminiModeOptions(approvalPolicy?: ApprovalPolicy): GeminiModeOption[] {
-  const options = GEMINI_MODE_SELECTION_OPTIONS as ReadonlyArray<{ id: string; label: string }>;
-  return options
-    .filter((option) => approvalPolicy === 'yolo' || option.id !== 'yolo')
-    .map((option, index) => ({
-      id: option.id,
-      shortLabel: option.label || deriveGeminiModeLabel(option.id),
-      badge: option.id === 'yolo'
-        ? '무승인'
-        : index === 0
-          ? '기본'
-          : '설정 가능',
-    }));
-}
-
-export function resolveComposerModels(
-  agent: AgentFlavor,
-  providerSelections?: ProviderModelSelections,
-  legacyCustomModels?: LegacyCustomModels,
-): ComposerModelOption[] {
-  const baseModels = isSupportedAgentFlavor(agent)
-    ? COMPOSER_MODELS_BY_AGENT[agent]
-    : COMPOSER_MODELS_BY_AGENT.codex;
-
-  const selectedModelIds = isSupportedAgentFlavor(agent)
-    ? (providerSelections?.[agent]?.selectedModelIds ?? [])
-    : [];
-  if (selectedModelIds.length > 0) {
-    return selectedModelIds.map((modelId: string, index: number) => {
-      const baseMatch = baseModels.find((item) => item.id === modelId);
-      if (baseMatch) {
-        return {
-          ...baseMatch,
-          badge: index === 0 ? '등록됨' : baseMatch.badge,
-        };
-      }
-      return {
-        id: modelId,
-        shortLabel: agent === 'codex' ? deriveOpenAiModelLabel(modelId) : modelId,
-        badge: index === 0 ? '등록됨' : '선택됨',
-      };
-    });
-  }
-
-  if (legacyCustomModels) {
-    const customId = legacyCustomModels[agent];
-    if (customId && customId.trim() !== '') {
-      const trimmed = customId.trim();
-      return [
-        { id: trimmed, shortLabel: trimmed, badge: '커스텀' },
-        ...baseModels.filter((model) => model.id !== trimmed),
-      ];
-    }
-  }
-  return baseModels;
-}
-
-export function resolveDefaultModelId(
-  agent: AgentFlavor,
-  providerSelections?: ProviderModelSelections,
-  legacyCustomModels?: LegacyCustomModels,
-  cachedModelId?: string | null,
-): string {
-  const availableModels = resolveComposerModels(agent, providerSelections, legacyCustomModels);
-  if (isSupportedAgentFlavor(agent)) {
-    const preferred = resolvePreferredModelId({
-      availableModelIds: availableModels.map((model) => model.id),
-      cachedModelId: agent === 'codex' ? cachedModelId : null,
-      configuredDefaultModelId: normalizeModelId(providerSelections?.[agent]?.defaultModelId),
-      fallbackModelId: 'gpt-5.4',
-    });
-    return preferred ?? 'gpt-5.4';
-  }
-  return availableModels[0]?.id ?? 'gpt-5.4';
-}
-
-function resolvePreferredModelId(input: {
-  availableModelIds: string[];
-  cachedModelId?: string | null;
-  configuredDefaultModelId?: string | null;
-  fallbackModelId: string;
-}): string | null {
-  const availableModelIds = new Set(input.availableModelIds);
-  if (input.cachedModelId && availableModelIds.has(input.cachedModelId)) {
-    return input.cachedModelId;
-  }
-  if (input.configuredDefaultModelId && availableModelIds.has(input.configuredDefaultModelId)) {
-    return input.configuredDefaultModelId;
-  }
-  if (availableModelIds.has(input.fallbackModelId)) {
-    return input.fallbackModelId;
-  }
-  return input.availableModelIds[0] ?? null;
-}
-
-export function resolveAvailableComposerModelId(input: {
-  agent: AgentFlavor;
-  requestedModel?: unknown;
-  sessionModelFallback?: unknown;
-  providerSelections?: ProviderModelSelections;
-  legacyCustomModels?: LegacyCustomModels;
-}): string {
-  const availableModels = resolveComposerModels(
-    input.agent,
-    input.providerSelections,
-    input.legacyCustomModels,
-  );
-  const availableIds = new Set(availableModels.map((model) => model.id));
-  const requestedModel = normalizeModelId(input.requestedModel);
-  if (requestedModel && availableIds.has(requestedModel)) {
-    return requestedModel;
-  }
-  const sessionModelFallback = normalizeModelId(input.sessionModelFallback);
-  if (sessionModelFallback && availableIds.has(sessionModelFallback)) {
-    return sessionModelFallback;
-  }
-  return availableModels[0]?.id ?? 'gpt-5.4';
-}
-
-export function resolveDefaultGeminiModeId(
-  approvalPolicy?: ApprovalPolicy,
-  configuredModeId?: unknown,
-): string {
-  const availableModes = resolveGeminiModeOptions(approvalPolicy);
-  if (approvalPolicy === 'yolo' && availableModes.some((mode) => mode.id === 'yolo')) {
-    return 'yolo';
-  }
-  const configuredMode = normalizeGeminiModeId(configuredModeId);
-  if (configuredMode && availableModes.some((mode) => mode.id === configuredMode)) {
-    return configuredMode;
-  }
-  return availableModes[0]?.id ?? DEFAULT_GEMINI_MODE_ID;
-}
-
-export function resolveAvailableGeminiModeId(input: {
-  requestedMode?: unknown;
-  approvalPolicy?: ApprovalPolicy;
-  configuredModeId?: unknown;
-}): string {
-  const availableModes = resolveGeminiModeOptions(input.approvalPolicy);
-  const availableIds = new Set(availableModes.map((mode) => mode.id));
-  const requestedMode = normalizeGeminiModeId(input.requestedMode);
-  if (requestedMode && availableIds.has(requestedMode)) {
-    if (requestedMode === 'yolo' && input.approvalPolicy !== 'yolo') {
-      return resolveDefaultGeminiModeId(input.approvalPolicy, input.configuredModeId);
-    }
-    return requestedMode;
-  }
-  return resolveDefaultGeminiModeId(input.approvalPolicy, input.configuredModeId);
 }
 
 export function isUserEvent(event: UiEvent): boolean {
